+++ src/phosphorus/__main__.py
@@ -1,5 +1,9 @@
 from phosphorus.lib.parser import parse_args
+from phosphorus.subcommands.build import BuildCommand
 
 
 def main() -> None:
-    parse_args()
+    args = parse_args()
+    match args.subcommand:
+        case "build":
+            BuildCommand(args).run()